dart: dcp/disp0: Set all pre-allocated L2 tables up on init
DART nodes for dcp and disp0 have pre-allocated L1 and L2 tables which
are annotated in the ADT. The pre-allocated memory is specified in
"pt-region-${DEVICE}". The first page is used as L1 table and the
following pages are used as L2 tables. The number of valid L2 tables is
specified in "l2-tt-${DEVICE}". The first entry identifies the region
and the second entry is the number of valid L2 tables.

dart-dcp and dart-disp0 have with macos 12.3 just 2 L2 tables setup.
This makes it impossible to use larger framebuffers than the
pre-allcoated framebuffer of 21MB. The additional L2 tables must remain
valid after m1n1 cedes control to the next target. This designated way
is the simplest solution and least likely to cause problems with macos.

Signed-off-by: Janne Grunau <j@jannau.net>